+++ lib/src/features/auth/services/bloc/states/auth_state.freezed.dart
@@ -55,6 +55,7 @@ mixin _$AuthState {
     required TResult Function(dynamic error, StackTrace stackTrace) error,
   }) =>
       throw _privateConstructorUsedError;
+
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult Function()? init,
@@ -64,6 +65,7 @@ mixin _$AuthState {
     TResult Function(dynamic error, StackTrace stackTrace)? error,
   }) =>
       throw _privateConstructorUsedError;
+
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? init,
@@ -74,6 +76,7 @@ mixin _$AuthState {
     required TResult orElse(),
   }) =>
       throw _privateConstructorUsedError;
+
   @optionalTypeArgs
   TResult map<TResult extends Object?>({
     required TResult Function(AuthStateInit value) init,
@@ -83,6 +86,7 @@ mixin _$AuthState {
     required TResult Function(AuthStateError value) error,
   }) =>
       throw _privateConstructorUsedError;
+
   @optionalTypeArgs
   TResult? mapOrNull<TResult extends Object?>({
     TResult Function(AuthStateInit value)? init,
@@ -92,6 +96,7 @@ mixin _$AuthState {
     TResult Function(AuthStateError value)? error,
   }) =>
       throw _privateConstructorUsedError;
+
   @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(AuthStateInit value)? init,
@@ -115,6 +120,7 @@ class _$AuthStateCopyWithImpl<$Res> implements $AuthStateCopyWith<$Res> {
   _$AuthStateCopyWithImpl(this._value, this._then);
 
   final AuthState _value;
+
   // ignore: unused_field
   final $Res Function(AuthState) _then;
 }
@@ -612,6 +618,7 @@ abstract class $AuthStateErrorCopyWith<$Res> {
   factory $AuthStateErrorCopyWith(
           AuthStateError value, $Res Function(AuthStateError) then) =
       _$AuthStateErrorCopyWithImpl<$Res>;
+
   $Res call({dynamic error, StackTrace stackTrace});
 }
 
@@ -766,7 +773,9 @@ abstract class AuthStateError implements AuthState {
 
 // ignore: avoid_annotating_with_dynamic
   dynamic get error;
+
   StackTrace get stackTrace;
+
   @JsonKey(ignore: true)
   $AuthStateErrorCopyWith<AuthStateError> get copyWith =>
       throw _privateConstructorUsedError;