+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+- feat: SignInScreenVm - add phone authorization
+
 ## 0.3.0
 
 - feat: vk login
+- feat: google sign in
 - feat: SignInScreenVm
 - fix: iOS build support
 - feat: auth repository